game 신청 취소 api 수정 및 테스트 코드 작성

+++ src/main/java/kr/couchcoding/tennis_together/domain/game/service/GameUserListService.java
@@ -61,14 +61,16 @@ public class GameUserListService {
                 .orElseThrow(() -> new CustomException(ErrorCode.BAD_REQUEST_GAME, "해당 게임에 신청한 이력이 없습니다."));
 
         LocalDate now = LocalDate.now();
-        if (((game.getStrDt().equals(now) || game.getEndDt().equals(now) ||
-                (game.getStrDt().isBefore(now) && game.getEndDt().isAfter(now))) &&
-                gameUserList.getStDvCd() == GameUserListStatus.APPROVED)) {
-            game.updateStatus(GameStatus.RECRUITING);
-        }
+        if (game.getStrDt().equals(now) || game.getEndDt().equals(now) ||
+                (game.getStrDt().isBefore(now) && game.getEndDt().isAfter(now))) {
+            if (gameUserList.getStDvCd() == GameUserListStatus.APPROVED)
+                game.updateStatus(GameStatus.RECRUITING);
+        } else
+            throw new CustomException(ErrorCode.BAD_REQUEST_GAME, "신청 기간이 지난 게임은 취소가 불가능 합니다.");
 
-        // 신청 승락 되기 전엔 자유롭게 취소 가능
-        // 근데 신청 승락이 되고 게임은 CLOSED가 되었을떄 취소한다면??? 미리 리뷰도 달았다면?
+        // 거절된 후 게임 요청 내역을 삭제하고 재신청을 방지하기 위함
+        if (gameUserList.getStDvCd() == GameUserListStatus.REFUSED)
+            throw new CustomException(ErrorCode.BAD_REQUEST_GAME, "거절된 게임의 신청을 취소할 수 없습니다.");
 
         gameUserListRepository.delete(gameUserList);
     }
